requirements.txt
certifi==2024.2.2
charset-normalizer==2.0.12
click==8.1.7
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
greenlet==3.0.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mysqlclient==2.2.4
numpy==1.26.4
pandas==1.3.3
PyMySQL==1.1.0
python-dateutil==2.8.2
pytz==2024.1
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.26
urllib3==1.26.18
pandas==1.3.3